backend/requirements-docker.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.2.0
python-multipart==0.0.6
azure-ai-inference==1.0.0b1
azure-identity>=1.12.0
openai>=1.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1